requirements.txt
streamlit==1.36.0
docxtpl==0.16.7
pypandoc==1.13
python-dotenv==1.0.1
python-docx>=1.1